move away from XPC
SourceKitten has previously always assumed that payloads were XPC objects.
This isn't always the case with SourceKit, which exposes the same interface
regardless of if you're using `sourcekitd.framework` or
`libsourcekitdInProc.dylib`.

This commits removes SourceKitten's dependency on SwiftXPC, which will be
fully removed in an upcoming commit.

This also has the nice advantage of avoiding the fragile parsing of binary
data for SyntaxMap and CodeCompletion.