Remove copy framework build phase for SourceKittenFramework target
The target is not depending on Commandant and Result, and the dependency
on SWXMLHash is already declared by the Link Binary with Libraries build
phase. This will allow the user of the framework to add SWXMLHash to
their dependency. And the copy framework action should happen when the
framework user target builds, like in the sourcekitten target.