Remove "key.offset" and "key.length" from results of `dictWithCommentMarkNamesCursorInfo()`
Their values are same with "key.nameoffset" and "key.namelength" in most case.
When kind is extension, their values locate **the type's declaration** in their declared file.
That may be different from the file declaring extension.

So, using them is wrong.
Is this related to https://github.com/realm/jazzy/issues/454?