Change completion's compilerargs to array of strings
Currently the compilerargs are consumed from commandant as a single
quoted string. Because of this, we have to manually split the string at
spaces, which causes issues with arguments, or filepaths with spaces in
them. Instead of doing this we can allow commandant to pass us the
"remaining" arguments trailing the `--` so we can avoid having to split
anything ourselves.