remove UID map and lock
The UID cache was added over 2 years ago in fe2cca6 and its impact was never
revisited until now.

Prompted by @merowing, I measured the impact of a few common operations with
and without the cache and determined that it should be removed, since it
hurts more than it helps.

Especially after the lock was added in #327. Even before the lock, the gains
were minimal.

SourceKitten unit tests

* with cache: 44.7s
* without cache: 45.8s

Vapor 1.3.8 with `doc --spm-module Vapor`

* with cache: 24.75s
* without cache: 21.52s

SwiftLint linting the Swift stdlib

* with cache: 43.7s
* without cache: 25.2s