Add package build step to SPM docs generation (#614)
Final part of jazzy+SPM support: this runs `swift build` as part of generating docs for SPM modules.  This makes sure that the build manifest (a) exists and (b) is consistent with the latest changes to the source code, same as Xcode path.  (No need to do a `clean` so is v. cheap if the package has already been built.)

First commit is a pure refactor of all the popen-ish code because I couldn't stand duplicating it again.  Accounts for most of the changed LOC.

Second commit adds a `Module` initializer that runs `swift build` before going to SourceKit.

I tentatively kept the existing SPM initializer to keep semantics for existing users (SourceDocs uses it, didn't look further).