Add InProc Option for Sourcekitd (#728)
There are two versions of sourcekitd that come with Xcode:

- `sourcekitd.framework`, which uses XPC to connect to the sourcekit daemon and parse swift files out-of-process.
- `sourcekitdInProc.framework`, which does the parsing in-process.

SourceKitten uses the XPC version on macOS. However, there may be cases, when the in-process version might be preferred. So this PR adds the possibility to configure this using the following environment variable:

    USE_INPROC_SOURCEKIT=YES

Specifically, this will be needed to support running SourceKitten in Swift Package Manager build tool plugins. It is a new feature ([SE-0303](https://github.com/apple/swift-evolution/blob/main/proposals/0303-swiftpm-extensible-build-tools.md)) in Swift 5.6, that allows Swift packages to execute custom build commands when building. So for example, you could have [swiftlint](https://github.com/realm/SwiftLint) run during a build of your Swift package.

As of this writing, Swift 5.6 is available in the Release Candidate of Xcode 13.3. I have experimented with it and found that SPM build commands run in a sandbox that prevent the use of XPC and crash when using the regular version of sourcekitd, but it works normally with sourcekitdInProc.