SourceKittenFramework: use `free` rather than `UMRP.deallocate`
The `UMRP.deallocate` will bridge to `_aligned_free` which requires that
the data be allocated with `_aligned_malloc`.  This does not hold as the
string is being duplicated via `strdup` and must be free'd explicitly
with `free`.  This happens to work on existing platforms as Darwin and
Linux do not use a separate memory arena for aligned memory allocations.