SourceKittenFramework: largely port library loading for Windows (#784)
This adds Windows specific handling for dynamic symbol resolution into
a library.  The `dlopen` and `dlsym` APIs are not portable as they are
part of POSIX rather than the C standard.  Use the Windows equivalents
of `LoadLibraryW` and `GetProcAddress`.

TODO: this requires further refinement to detect the Swift toolchain
installation.